pyproject.toml
[project]
name = ""
version = "0.0.1"
requires-python = "~=3.13"
dependencies = [
    "django",
    "django-taggit",
    "django-imagekit",
    "pillow",
    "lxml",
    "fontawesomefree",
    "django-crispy-forms",
    "crispy-bulma",
    "gunicorn",
    "django-select2",
    "django-meta",
    "psycopg[binary]",
    "django-ckeditor-5>=0.2.17",
    "django-ckeditor>=6.7.2",
    "django-imagefield>=0.22.0",
]

[tool.django-stubs]
# Point this to your settings module, e.g. "myproject.settings"
django_settings_module = "backend/blogsite.settings"

[dependency-groups]
dev = [
    "django-debug-toolbar",
    "djlint>=1.36.4",
    "ipython",
]

[tool.djlint]
indent = 2

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.ruff]
extend-exclude = ["migrations"]
extend-select = ["A", "B", "C4", "C90", "COM", "DJ", "I", "LOG", "N", "PTH", "RUF", "SIM", "W"]
line-length = 119
ignore = ["RUF012"]
